feat(sqp): Support metrics chunk
Add support for SQP metrics chunk. Introduces a new method, `SetMetric(index byte, value float32)` on `Server` to enable SDK users to set arbitrary metrics up to the pre-defined limit.

Also, address a few comments from the previous PR.

+++ game-server-hosting/server/proto/sqp/errors.go
@@ -8,7 +8,7 @@ import (
 type (
 	// UnsupportedSQPVersionError is an error which represents an invalid SQP version provided to the reader.
 	UnsupportedSQPVersionError struct {
-		version int8
+		version uint16
 	}
 )
 
@@ -18,7 +18,7 @@ var (
 )
 
 // NewUnsupportedSQPVersionError returns a new instance of UnsupportedSQPVersionError.
-func NewUnsupportedSQPVersionError(version int8) error {
+func NewUnsupportedSQPVersionError(version uint16) error {
 	return &UnsupportedSQPVersionError{
 		version: version,
 	}
